add correct type to "norm by max"

+++ +Calibration/+aux/+CBTools/checkerboardInfoMessage.m
@@ -1,6 +1,6 @@
 function [  ] = checkerboardInfoMessage( frame,fprintff ,th )
 warning('off','vision:calibrate:boardShouldBeAsymmetric'); % Supress checkerboard warning
-[p,~] = detectCheckerboardPoints(normByMax(frame.i)); % p - 3 checkerboard points. bsz - checkerboard dimensions.
+[p,~] = detectCheckerboardPoints(normByMax(double(frame.i))); % p - 3 checkerboard points. bsz - checkerboard dimensions.
 if size(p,1)<th
     fprintff('%d/%d CB corners detected.\n',size(p,1),th);
 end
